Nested NextJS routing! Migrate to Next.js 14 route-based layouts. This commit restructures app layout to properly utilize Next.js 14's built-in layout system, fixing issues with missing headers and navbars. Moved user layout from components folder to src/app/user/layout.js while preserving UserLayout component name for codebase clarity. Moved admin layout from components folder to src/app/admin/layout.js while preserving AdminLayout component name, Updated import paths in layouts to properly reference components. Fixed React params handling in profile pages using React.use(). Moved console.logs to useEffect to prevent render-time side effects. Maintained existing auth logic and conditional rendering in layouts. Kept UnauthenticatedLayout in components folder for reuse (FOR NOW). ALL layout files must be named 'layout.js' for Next.js to recognize them. Layouts now properly wrap their respective route segments. Auth state and user checks remain at layout level. Profile pages now properly resolve params using React.use().
