+++ src/app/admin/layout.js
@@ -1,11 +1,11 @@
 "use client";
 
-import UserHeader from "./UserHeader";
-import UserNavbar from "./UserNavbar";
-import UnauthenticatedLayout from "./UnauthenticatedLayout";
+import AdminHeader from "../components/AdminHeader";
+import AdminNavbar from "../components/AdminNavbar";
+import UnauthenticatedLayout from "../components/UnauthenticatedLayout";
 import { useSelector } from "react-redux";
 
-const UserLayout = ({ children }) => {
+const AdminLayout = ({ children }) => {
   const { user } = useSelector((state) => state.auth);
 
   if (!user) {
@@ -14,13 +14,13 @@ const UserLayout = ({ children }) => {
 
   return (
     <div className="d-flex flex-column min-vh-100">
-      <UserHeader />
+      <AdminHeader />
       <main className="flex-grow-1">
         {children}
       </main>
-      <UserNavbar />
+      <AdminNavbar />
     </div>
   );
 };
 
-export default UserLayout;
+export default AdminLayout;
