+++ src/app/page.js
@@ -1,22 +1,28 @@
 "use client";
 
+import { useEffect } from 'react';
 import { useSelector } from "react-redux";
+import { useRouter } from 'next/navigation';
 import dynamic from "next/dynamic";
 
 const GuestHome = dynamic(() => import("./GuestHome"), {
   loading: () => <div>Loading GuestHome component...</div>
 });
 
-const UserPage = dynamic(() => import("./user/[id]/profile/page"), {
-  loading: () => <div>Loading UserPage component...</div>
-});
-
-const AdminPage = dynamic(() => import("./admin/[id]/profile/page"), {
-  loading: () => <div>Loading AdminPage component...</div>
-});
-
 export default function Root() {
+  const router = useRouter();
   const { user, role, isInitialized } = useSelector((state) => state.auth);
+
+  useEffect(() => {
+    if (!isInitialized) return;
+
+    if (user) {
+      const path = role === 'admin' ? 
+        `/admin/${user.id}/profile` : 
+        `/user/${user.id}/profile`;
+      router.push(path);
+    }
+  }, [user, role, isInitialized, router]);
 
   // Wait for auth to be initialized
   if (!isInitialized) {
@@ -28,8 +34,6 @@ export default function Root() {
     return <GuestHome />;
   }
 
-  // Show admin or user page based on role
-  return role === 'admin' ?
-    <AdminPage params={{ id: user.id }} /> :
-    <UserPage params={{ id: user.id }} />;
+  // Show loading while redirect happens
+  return <div>Redirecting to dashboard...</div>;
 }
